requirements.txt
alembic==1.9.4
anyio==3.6.2
async-timeout==4.0.2
beautifulsoup4==4.11.2
bs4==0.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
JSON-log-formatter==0.5.1
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.5
python-dotenv==0.21.1
redis==4.5.1
schedule==1.1.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
